FaceRecognitionApp/requirements.txt
keras~=2.10.0
scipy~=1.9.3
opencv-python~=4.6.0.66
torchvision~=0.13.1
torch~=1.13.0
tensorflow
numpy~=1.23.5
pillow~=9.3.0